Make sure /var/log/journal exists.
This is necessary because on systems (Ubuntu 16.04) that have both
`Storage=auto` and `SplitMode=uid` may *NOT* have `journalctl --user`
functional.

SplitMode=uid only works if the storage is persistent (according to
journald's man page), and `Storage=auto` will operate as volatile
(instead of persistent) if there is no /var/log/journal directory. This
has the effect of Ubuntu 16.04 shipping with volatile logs *and* no
user journals.

I like `journalctl --user` to work, hence this patch.

This should solve the following problem:

    % journalctl --user
    No journal files were opened due to insufficient permissions.